refactor(runtime): security hardening, thread safety, and cleanup
- Restore GPIO capability guards (critical security regression fix)
- Gate mem_alloc/free on new AKIRA_CAP_MEMORY capability
- Route bt_adv start/stop through AKIRA_CHECK_CAP_OR_RETURN for audit log
- Split AKIRA_CAP_APP_INFO (read) from AKIRA_CAP_APP_CONTROL (write)
- Fix NULL return in akira_capability_mask_to_str() → "unknown"
- Fix sensor read returning INT32_MIN on denial → -EACCES
- Replace shared g_exit_mutex with per-slot exit_mutex in akira_managed_app_t
- Make g_runtime_initialized atomic_t with atomic_cas init guard
- Lock g_runtime_mutex in get_name_for_module_inst during linear scan
- Protect g_exit_cb reads/writes under g_runtime_mutex
- Add s_rf_lock mutex to all RF API functions; fix deinit double-lock
- Switch sandbox rate bucket timestamp to uint32_t + k_uptime_get_32()
  for native atomicity on 32-bit Xtensa
- Replace k_malloc/k_free with akira_malloc_buffer/akira_free_buffer in
  app_loader and app_manager WASM staging buffers
- Extract K_SECONDS(15/5/3) magic numbers into Kconfig symbols
- Remove unregistered akira_display_module.c stub from build
- Wire up akira_register_log_module/time_module in runtime init
- Fix LOG_MODULE_REGISTER to use CONFIG_AKIRA_LOG_LEVEL (6 files)

+++ src/runtime/security/sandbox.h
@@ -60,7 +60,7 @@ typedef struct {
     atomic_t tokens;            /**< Current tokens */
     uint16_t max_tokens;        /**< Maximum tokens (burst capacity) */
     uint16_t refill_per_sec;    /**< Tokens refilled per second */
-    int64_t last_refill_ms;     /**< Last refill timestamp */
+    uint32_t last_refill_ms;    /**< Last refill timestamp (32-bit, natively atomic on Xtensa) */
 } sandbox_rate_bucket_t;
 
 /** Rate limit configuration */
